move common tailFile to a separate file
