+++ lib/winston/transports/file.js
@@ -17,7 +17,7 @@ const TransportStream = require('winston-transport');
 const PassThrough = require('stream').PassThrough;
 const debug = require('diagnostics')('winston:file');
 const os = require('os');
-const common = require('../common');
+const tailFile = require('../tail-file');
 
 var noop = function () {};
 
@@ -286,7 +286,7 @@ File.prototype.stream = function (options) {
     start: options.start
   };
 
-  stream.destroy = common.tailFile(tail, function (err, line) {
+  stream.destroy = tailFile(tail, function (err, line) {
     if (err) {
       return stream.emit('error', err);
     }
